Implement backend user authentication.
Email now associated with user at create time and used for login.

Argon2id with per-user salt is used for key derivation - key and
salt stored on the user record - password is never kept.

JWT authorization implemented. User ID is encoded in token
generated at login. This is parsed by middleware for each request
if present and stored in the context for use by GraphQL resolvers.

Where current user is required this is now injected from the user
auth. If authorization is required e.g. to create/delete rules then
this is now enforced.